+++ src/sentry/shared_integrations/client/__init__.py
@@ -1,0 +1,8 @@
+__all__ = (
+    "BaseApiClient",
+    "BaseInternalApiClient",
+    "BaseApiResponse",
+)
+
+from .base import BaseApiClient, BaseApiResponse
+from .internal import BaseInternalApiClient
