+++ src/sentry/shared_integrations/response/mapping.py
@@ -1,0 +1,11 @@
+from sentry.shared_integrations.response.base import BaseApiResponse
+
+
+class MappingApiResponse(dict, BaseApiResponse):
+    def __init__(self, data, *args, **kwargs):
+        dict.__init__(self, data)
+        BaseApiResponse.__init__(self, *args, **kwargs)
+
+    @property
+    def json(self):
+        return self
