+++ src/sentry/shared_integrations/response/xml.py
@@ -1,0 +1,9 @@
+from bs4 import BeautifulSoup
+
+from sentry.shared_integrations.response.base import BaseApiResponse
+
+
+class XmlApiResponse(BaseApiResponse):
+    def __init__(self, text, *args, **kwargs):
+        self.xml = BeautifulSoup(text, "xml")
+        super().__init__(*args, **kwargs)
